[0.6.2.198] - Fix nelements FocusOn/Out Notification v2, have 2 bugs: 1)if fast close notification -> see it 2) Before 9 notification (included) text wasn't rendering

+++ src/game/client/components/notification.cpp
@@ -1,5 +1,6 @@
 #include <base/vmath.h>
 #include <engine/graphics.h>
+#include <engine/shared/config.h>
 #include "notification.h"
 
 CNotification::CNotification()
@@ -11,9 +12,9 @@ CNotification::CNotification()
 
 void NotifyAnimationDie(CNUIElements *Main, void *arg)
 {
-	CNotification *pThis = (CNotification *)arg;
+	//CNotification *pThis = (CNotification *)arg;
 	Main->SetLifeTime(0, 1);
-	Main->SetEndLifeAnimation(EaseIN6, vec4(210, 0, -140, 0));
+	Main->SetEndLifeAnimation(EaseIN6, vec4(210, 0, 0, 0));
 }
 
 void NotifyDie(CNUIElements *Main, void *arg)
@@ -23,36 +24,51 @@ void NotifyDie(CNUIElements *Main, void *arg)
 	pThis->GetNotifyArray()->remove(Main);
 
 	if(pThis->GetNotifyWaitArray()->size() > 0)
+	{
 		pThis->RenderNotification(pThis->GetNextNotify()->Type, pThis->GetNextNotify()->pTitle, pThis->GetNextNotify()->pText);
+		pThis->GetNotifyWaitArray()->remove_index(0);
+	}
+	pThis->SetSize(pThis->GetNotifyWaitArray()->size());
+	if(pThis->GetNotifyArray()->size() == 0)
+	{
+		CNUIElements *NotifyBox = pThis->ControllerNUI()->GetElement(ELEMENT_BLOCK, "Notification.NotifyBox");
+		NotifyBox->GetPos()->Init(pThis->GetNotifyBoxStartPos());
+	}
+}
 
-	pThis->GetNotifyWaitArray()->remove_index(0);
+void NotifyBoxAnimEnd(CNUIElements *Main, void *arg)
+{
+	CNotification *pThis = (CNotification *)arg;
+	if(pThis->GetNotifyArray()->size() < 5 && pThis->GetNotifyWaitArray()->size() > 0)
+	{
+		if(pThis->GetNotifyWaitArray()->size() > 0)
+			pThis->RenderNotification(pThis->GetNextNotify()->Type, pThis->GetNextNotify()->pTitle, pThis->GetNextNotify()->pText);
+
+		pThis->GetNotifyWaitArray()->remove_index(0);
+	}
+	pThis->SetSize(pThis->GetNotifyWaitArray()->size());
 }
 
 bool CNotification::AddNew(NotificationType Type, const char *pTitle, const char *pText)
 {
 	if(m_aNotify.size() < 5)
 	{
-
 		CNUIElements *NotifyBox = ControllerNui()->GetElement(ELEMENT_BLOCK, "Notification.NotifyBox");
-		NotifyBox->GetPos()->InitPlus(vec4(0, -90, 0, 90), 1, EaseOUT4);
+		NotifyBox->GetPos()->InitPlus(vec4(0, -70, 0, 70), 1, EaseOUT4);
+		NotifyBox->SetAnimationEndCallback(NotifyBoxAnimEnd, this);
+
 		TempNames[m_NowRendering] = new char[64] {0};
 		str_format(TempNames[m_NowRendering], 64, "Notification.NotifyBox.Notify%d", m_NowRendering);
-		//CUIRect Screen, MainRect, TextRect;
 
 		CNUIElements *Main = ControllerNui()->GetElement(ELEMENT_BLOCK, TempNames[m_NowRendering]);
-/*		Graphics()->GetScreen(&Screen.x, &Screen.y, &Screen.w, &Screen.h);
-		Screen.VSplitRight(50.0f, 0, &MainRect);
-		MainRect.HSplitBottom(175.0f, &MainRect, 0);
-		MainRect.HSplitBottom(50.0f, 0, &MainRect);*/
 
-		Main->GetPos()->Init(vec4(NotifyBox->GetPos()->m_Value.w+70, NotifyBox->GetPos()->m_Value.h, NotifyBox->GetPos()->m_Value.w, 70));
-		//Main->GetPos()->Init(vec4(0, 0, NotifyBox->GetPos()->m_Value.w, NotifyBox->GetPos()->m_Value.h));
+		Main->GetPos()->Init(vec4(NotifyBox->GetPos()->m_Value.w+70, NotifyBox->GetPos()->m_Value.h, NotifyBox->GetPos()->m_Value.w, 60));
 		Main->GetPos()->InitPlus(vec4(-(NotifyBox->GetPos()->m_Value.w+70), 0, 0, 0), 1, EaseOUTBounce);
 		Main->SetLifeTime(10, 1);
 		Main->SetBlock(5.0f, CORNER_ALL);
 		Main->GetColor()->Init(vec4(0, 1, 0, 0.4f));
-		//Main->GetPos()->InitPlus(vec4(-210, 0, 140, 20), 1, EaseOUT5);
 		Main->SetEndLifeTimeCallback(NotifyAnimationDie, NotifyDie, this);
+		Main->SetCallbacksEvents(NotifyAnimationDie, NULL, NULL, this);
 
 		TempTitle[m_NowRendering] = new char[64] {0};
 		str_format(TempTitle[m_NowRendering], 64, "%s.Title", TempNames[m_NowRendering]);
@@ -99,6 +115,7 @@ void CNotification::Add(NotificationType Type, const char *pTitle, const char *p
 	Notify.pTitle = pTitle;
 	Notify.pText = pText;
 	m_aNotifyWait.add(Notify);
+	SetSize(GetNotifyWaitArray()->size());
 }
 
 void MainMouseOver(CNUIElements *Main, void *arg)
@@ -107,7 +124,7 @@ void MainMouseOver(CNUIElements *Main, void *arg)
 	Main->GetPos()->InitPlus(vec4(-10, -2, 0, 4), 1, EaseOUTElastic);
 	Main->GetColor()->InitPlus(vec4(0, 0, 0, 0.3f), 1, EaseOUTElastic);
 
-	CNUIElements *Text = pThis->ControllerNUI()->GetElement(ELEMENT_TEXT, "Notification.Box.Background.Text");
+	CNUIElements *Text = pThis->ControllerNUI()->GetElement(ELEMENT_TEXT, "Notification1.Background.Text");
 	Text->GetColor()->InitPlus(vec4(0,0,0,0.5f), 1, EaseOUTElastic);
 }
 
@@ -117,7 +134,7 @@ void MainMouseOut(CNUIElements *Main, void *arg)
 	Main->GetPos()->InitPlus(vec4(10, 2, 0, -4), 1, EaseOUTElastic);
 	Main->GetColor()->InitPlus(vec4(0, 0, 0, -0.3f), 1, EaseOUTElastic);
 
-	CNUIElements *Text = pThis->ControllerNUI()->GetElement(ELEMENT_TEXT, "Notification.Box.Background.Text");
+	CNUIElements *Text = pThis->ControllerNUI()->GetElement(ELEMENT_TEXT, "Notification1.Background.Text");
 	Text->GetColor()->InitPlus(vec4(0,0,0,-0.5f), 1, EaseOUTElastic);
 }
 
@@ -125,12 +142,9 @@ void MainClick(CNUIElements *Main, void *arg)
 {
 	CNotification *pThis = (CNotification *)arg;
 
-	for(int i = 0; i < pThis->GetNotifyWaitArray()->size(); i++)
+	if(pThis->GetNotifyWaitArray()->size() > 0)
 	{
-		if(i==5)
-			break;
 		pThis->RenderNotification(pThis->GetNextNotify()->Type, pThis->GetNextNotify()->pTitle, pThis->GetNextNotify()->pText);
-
 		pThis->GetNotifyWaitArray()->remove_index(0);
 	}
 }
@@ -140,7 +154,7 @@ void CNotification::OnRender()
 	if(!m_Init)
 	{
 		m_Size = 0;
-		CNUIElements *Main = ControllerNui()->GetElement(ELEMENT_BLOCK, "Notification.Box.Background");
+		CNUIElements *Main2 = ControllerNui()->GetElement(ELEMENT_BLOCK, "Notification1.Background");
 
 		CUIRect Screen, MainRect, TextRect;
 		Graphics()->GetScreen(&Screen.x, &Screen.y, &Screen.w, &Screen.h);
@@ -149,22 +163,24 @@ void CNotification::OnRender()
 		MainRect.HSplitBottom(175.0f, &MainRect, 0);
 		MainRect.HSplitBottom(50.0f, 0, &MainRect);
 
-		Main->GetPos()->Init(vec4(MainRect.x+MainRect.w, MainRect.y, MainRect.w, MainRect.h));
-		Main->GetPos()->Init(vec4(MainRect.x+10, MainRect.y, MainRect.w, MainRect.h), 1, EaseOUTElastic);
-		Main->SetBlock(5.0f, CORNER_L);
-		Main->GetColor()->Init(vec4(0, 1, 0, 0.2f));
-		Main->SetCallbacksVisual(MainMouseOver, MainMouseOut, NULL, NULL, this);
-		Main->SetCallbacksEvents(MainClick, NULL, NULL, this);
+		Main2->GetPos()->Init(vec4(MainRect.x+MainRect.w, MainRect.y, MainRect.w, MainRect.h));
+		Main2->GetPos()->Init(vec4(MainRect.x+10, MainRect.y, MainRect.w, MainRect.h), 1, EaseOUTElastic);
+		Main2->SetBlock(5.0f, CORNER_L);
+		Main2->GetColor()->Init(vec4(0, 1, 0, 0.2f));
+		Main2->SetCallbacksVisual(MainMouseOver, MainMouseOut, NULL, NULL, this);
+		Main2->SetCallbacksEvents(MainClick, NULL, NULL, this);
 
-		CNUIElements *Text = ControllerNui()->GetElement(ELEMENT_TEXT, "Notification.Box.Background.Text");
+		CNUIElements *Text = ControllerNui()->GetElement(ELEMENT_TEXT, "Notification1.Background.Text");
 		Text->GetPos()->Init(vec4(0, 15, 50, 20.0f));
 		Text->GetColor()->Init(vec4(1,1,1,0.5f));
 		Text->SetText(true, ALIGN_CENTER, "%d", &m_Size);
 
 
+		m_NotifyBoxStartPos = vec4(MainRect.x-210, MainRect.y, MainRect.w+140, MainRect.h+20);
 		CNUIElements *NotifyBox = ControllerNui()->GetElement(ELEMENT_BLOCK, "Notification.NotifyBox");
-		NotifyBox->GetPos()->Init(vec4(MainRect.x-210, MainRect.y, MainRect.w+140, MainRect.h+20));
-		NotifyBox->GetColor()->Init(vec4(0, 0, 0, 0.2f));
+		NotifyBox->GetPos()->Init(m_NotifyBoxStartPos);
+		if(g_Config.m_Debug)
+			NotifyBox->GetColor()->Init(vec4(0, 0, 0, 0.2f));
 
 		m_Init = true;
 	}
